+++ demos/browser/app/meeting/meeting.ts
@@ -134,6 +134,7 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver 
 
   // feature flags
   enableWebAudio = false;
+  enableUnifiedPlanForChromiumBasedBrowsers = true;
 
   constructor() {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -582,6 +583,7 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver 
     }
     const deviceController = new DefaultDeviceController(logger);
     configuration.enableWebAudio = this.enableWebAudio;
+    configuration.enableUnifiedPlanForChromiumBasedBrowsers = this.enableUnifiedPlanForChromiumBasedBrowsers;
     this.meetingSession = new DefaultMeetingSession(configuration, logger, deviceController);
     this.audioVideo = this.meetingSession.audioVideo;