+++ demos/browser/app/meetingV2/meetingV2.ts
@@ -141,6 +141,7 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver,
 
   // feature flags
   enableWebAudio = false;
+  enableUnifiedPlanForChromiumBasedBrowsers = true;
 
   constructor() {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -550,6 +551,7 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver,
     }
     const deviceController = new DefaultDeviceController(logger);
     configuration.enableWebAudio = this.enableWebAudio;
+    configuration.enableUnifiedPlanForChromiumBasedBrowsers = this.enableUnifiedPlanForChromiumBasedBrowsers;
     this.meetingSession = new DefaultMeetingSession(configuration, logger, deviceController);
     this.audioVideo = this.meetingSession.audioVideo;